+++ aliding-20230426/alibabacloud_aliding20230426/models/_create_run_request.py
@@ -12,17 +12,20 @@ class CreateRunRequest(DaraModel):
         assistant_id: str = None,
         ext_login_user: main_models.CreateRunRequestExtLoginUser = None,
         original_assistant_id: str = None,
+        source_id_of_assistant_id: str = None,
         source_id_of_original_assistant_id: str = None,
-        source_type_of_original_assistant_id: str = None,
+        source_type_of_assistant_id: int = None,
+        source_type_of_original_assistant_id: int = None,
         stream: bool = None,
         thread_id: str = None,
     ):
         self.allow_struct_view_content = allow_struct_view_content
-        # This parameter is required.
         self.assistant_id = assistant_id
         self.ext_login_user = ext_login_user
         self.original_assistant_id = original_assistant_id
+        self.source_id_of_assistant_id = source_id_of_assistant_id
         self.source_id_of_original_assistant_id = source_id_of_original_assistant_id
+        self.source_type_of_assistant_id = source_type_of_assistant_id
         self.source_type_of_original_assistant_id = source_type_of_original_assistant_id
         self.stream = stream
         # This parameter is required.
@@ -49,8 +52,14 @@ class CreateRunRequest(DaraModel):
         if self.original_assistant_id is not None:
             result['originalAssistantId'] = self.original_assistant_id
 
+        if self.source_id_of_assistant_id is not None:
+            result['sourceIdOfAssistantId'] = self.source_id_of_assistant_id
+
         if self.source_id_of_original_assistant_id is not None:
             result['sourceIdOfOriginalAssistantId'] = self.source_id_of_original_assistant_id
+
+        if self.source_type_of_assistant_id is not None:
+            result['sourceTypeOfAssistantId'] = self.source_type_of_assistant_id
 
         if self.source_type_of_original_assistant_id is not None:
             result['sourceTypeOfOriginalAssistantId'] = self.source_type_of_original_assistant_id
@@ -78,8 +87,14 @@ class CreateRunRequest(DaraModel):
         if m.get('originalAssistantId') is not None:
             self.original_assistant_id = m.get('originalAssistantId')
 
+        if m.get('sourceIdOfAssistantId') is not None:
+            self.source_id_of_assistant_id = m.get('sourceIdOfAssistantId')
+
         if m.get('sourceIdOfOriginalAssistantId') is not None:
             self.source_id_of_original_assistant_id = m.get('sourceIdOfOriginalAssistantId')
+
+        if m.get('sourceTypeOfAssistantId') is not None:
+            self.source_type_of_assistant_id = m.get('sourceTypeOfAssistantId')
 
         if m.get('sourceTypeOfOriginalAssistantId') is not None:
             self.source_type_of_original_assistant_id = m.get('sourceTypeOfOriginalAssistantId')